Simulate Datasets

Simulates datasets using R or JAGS code. By defaults return the datasets as an nlists_object. If path is provided then the datasets are written to the directory as individual .rds files.

sims_simulate(code, constants = nlist::nlist(),
  parameters = nlist::nlist(), monitor = ".*", stochastic = TRUE,
  latent = FALSE, nsims = getOption("sims.nsims", 100L),
  path = getOption("sims.path"), exists = FALSE,
  rdists = sims_rdists(), progress = FALSE,
  options = furrr::future_options(), ask = getOption("sims.ask", TRUE),
  silent = FALSE)

Arguments

code
A string of the JAGS code to generate the data. The JAGS code must not be in a data or model block.

constants
An nlist object (or list that can be coerced to nlist) specifying the values of nodes in code. The values are included in the output dataset.

parameters
An nlist object (or list that can be coerced to nlist) specifying the values of nodes in code. The values are not included in the output dataset.

monitor
A character vector (or regular expression if a string) specifying the names of the nodes in code to include in the dataset. By default all nodes are included.

stochastic
A logical scalar specifying whether to monitor deterministic and stochastic (NA), only deterministic (FALSE) or only stochastic nodes (TRUE).

latent
A logical scalar specifying whether to monitor observed and latent (NA), only latent (TRUE) or only observed nodes (FALSE).

nsims
An integer between 1 and 1,000,000 specifying the number of data sets to simulate. By default 100 data sets are simulated.

path
A string specifying the path to the directory to save the data sets in. By default path = NULL the data sets are not saved but are returned as an nlists object.

exists
A flag specifying whether the path directory should already exist. If exists = NA it doesn't matter. If the directory already exists all sims compatible files are deleted. otherwise an error is thrown.

rdists
A character vector specifying the R functions to recognize as stochastic.

progress
A flag specifying whether to print a progress bar.

options
The future specific options to use with the workers.

ask
A flag specifying whether to ask before deleting all sims compatible files.

silent
A flag specifying whether to suppress warnings.

Value

By default an nlists_object of the simulated data. Otherwise if path it returns TRUE.

Details

JAGS code is identified by the presence of '~' indicating a stochastic variable node. Otherwise code is assumed to be R code and stochastic variable nodes are those where assignment is immediately succeeded by a call to one of the functions named in rdists.

Both constants and parameters must be nlist_objects (or lists that can be coerced to such) . The only difference between constants and parameters is that the values in constants are appended to the output data while the values in parameters are not. Neither constants or parameters can include missing values nor can they have elements with the same name. Elements which are not in code are dropped with a warning (unless silent = TRUE in which case the warning is suppressed).

Each set of simulated data set is written as a separate .rds file. The files are labelled data0000001.rds, data0000002.rds, data0000003.rds etc. The argument values are saved in the hidden file .sims.rds.

sims compatible files are those matching the regular expression "^((data\\d{7,7})|([.]sims))[.]rds$".

Parallelization is accomplished using the future package. The progress and options arguments are both passed to future_map().

See also

Examples

set.seed(101)
sims_simulate("a ~ dunif(0, 1)", path = tempdir(), exists = NA, ask = FALSE)

#> Warning: Deleted 100 sims data files in '/var/folders/48/q6ltldjs251000_wvjrdy_vm0000gn/T//Rtmpelro2O'.

#> [1] TRUE