Simulate Datasets

Simulates datasets using JAGS code. By defaults return the datasets as an nlists_object. If path is provided then the datasets are written to the directory as individual .rds files.

sims_simulate(code, constants = nlist::nlist(),
  parameters = nlist::nlist(), monitor = ".*",
  nsims = getOption("sims.nsims", 100L), seed = rcount(),
  parallel = FALSE, path = NULL, exists = FALSE, silent = FALSE)

Arguments

code
A string of the JAGS code to generate the data. The code must not be in a data or model block.

constants
An nlist object specifying the values of nodes in code. The values are included in the output data.

parameters
An nlist object specifying the values of nodes in code. The values are not included in the output data.

monitor
A character vector (or regular expression if a string) specifying the names of the stochastic nodes in code to include in the data. By default all stochastic nodes are included.

nsims
An integer between 1 and 1,000,000 specifying the number of data sets to simulate. By default 100 data sets are simulated.

seed
A positive integer specifying the random seed to use for simulating the data.

parallel
A flag specifying whether to generate the datasets in parallel.

path
A string specifying the path to the directory to save the data sets in. By default path = NULL the data sets are not saved but are returned as an nlists object.

exists
A flag specifying whether the directory should already exist. If exists = NA it doesn't matter. If the directory already exists it is overwritten if exists = TRUE or exists = NA otherwise an error is thrown.

silent
A flag specifying whether to suppress warnings.

Value

By default an nlists_object of the simulated data. Otherwise if path is defined the simulated data argument values.

Details

Both constants and parameters must be nlist_objects or uniquely named lists of numeric vectors, matrices and arrays that can be coerced to nlist objects. The only difference between constants and parameters is that the values in constants are appended to the output data while the values in parameters are not. Neither constants or parameters can include missing values nor can they have elements with the same name. Elements which are not in code are dropped with a warning (unless silent = TRUE in which case the warning is suppressed).

Each set of simulated data set is written as a separate .rds file. The files are labelled data0000001.rds, data0000002.rds, data0000003.rds etc.

Examples

set.seed(101)
sims_simulate("a ~ dunif(0, 1)", path = tempdir(), exists = NA)

#> $code
#> [1] "model{a ~ dunif(0, 1)}\n"
#> 
#> $constants
#> an nlist object with 0 natomic elements
#> 
#> $parameters
#> an nlist object with 0 natomic elements
#> 
#> $monitor
#> [1] "a"
#> 
#> $nsims
#> [1] 100
#> 
#> $seed
#> [1] 799289926
#>